Sets CI/CD for the Music Store (#2)
TL;DR
-----

Adds a Concourse pipeline and scripts to simplify configuring it

Details
-------

Supports CI/CD for the Music Store demo application by adding a
Concourse pipeline and some scripts to simplify configuring it. The
pipeline executes build and test steps using .NET Core tools and 
then uses the Tanzu Build Service to create container images for 
each application/service that is part of the Music Store. It also uses 
`conftest` to validate the Kubernetes objects before deployment.

Has two scripts to facilitate setup:

1. `add-team.sh` which sets up Okta, Concourse, and Kubernetes
   for a new team workspace. The script creates and Okta group for
   the team, adds the required namespaces to the shared services 
   and workload clusters, and assure that the Okta group has the 
   right access to their namespaces.
2. `set-pipeline.sh` creates the Concourse pipeline. It's fairly
   straightforward right now but reduces the risk of typos.
   
The pipeline currently as a placeholder for static code scanning,
which is the next thing I'll be adding. It also has a commented 
out segment from its Java counterpart for doing an application 
scan that will come after that.