Automates Harbor configuration (#3)
TL;DR
-----

Scripts Harbor setup to make it easier and more repeatable

Details
-------

Adds a simple script that configure Harbor for the Lab. The script
sets up the projects, remote registries, and replications used by
the lab. Right now it doesn't check if the entities exist before
attempting to create them, so if you run it multiple times you'll
see 409 errors for the objects that have already been created.

Also includes two additiona replication configurations that are
needed for the Music Store infrastructure: the Spring Cloud Config
Server and Microsoft's SQL tools image which is useful for taking
a look at the SQL Server database if things are amiss.