Adds static scan to the pipeline (#5)
TL;DR
-----

Includes a real static scan in the pipeline

Details
-------

Updates the pipeline to use Sonarqube for a static code scan
instead of faking it out. To do this, I also added an image
bases on Microsofts .NET Core image that includes the Sonqarqube
plugin for the `dotnet` CLI and the JDK that it needs to run.

Includes a convenience script `build-and-push-scanner-image.sh`
that builds the image and pushes it to the private registry.